Data extracted from BGG 2018_09_12
Ranked games extracted from BoardGameGeek (BGG) using nanDeck tool. Because of the data size, data was extracted over 4 sessions of 4,000 records per session over the period of 2018-09-12 to 2018_09_13 - for a total of 16,000 records. Only core games are included in this dataset - no game expansions.
After extraction, the non-ranked games were deleted and the session records collated into one file saved as various file formats - CSV and XLSX files. After collation, there are 15,869 records - the highest ranked game being 1, the lowest 15,900. There are a few occurrences of games with identical ranks in this dataset.